Add support for the environment variables TEAM4121HOME and TEAM4121CONFIG.
The environment variable TEAM4121HOME allows the directory-of-control for
Team 4121 Vision Processing to be precisely specified before launch of
any vision-processing application.  This directory is the home location
for library files, configuration files, and logs.  By default, it will
take the value of the directory from which the Vision application was
launched.

The environment variable TEAM4121CONFIG provides a mechanism to precisely
select the set of Vision configuration files to be used from the
directory ${TEAM4121HOME}/config.  The default for this varies by
application.

All Python applications are updated to support launching as direct
shell applications, without needing to be prefixed with 'python3'.