Add support for running Team 4121 Vision applications on a PC.
Running Vision applications on a PC is facilitated by creation of
a 'docker' sandbox which emulates the environment of a Raspberry Pi.

Detailed directions for this are provided in the file 'info/README_pc.md',
which can be viewed with any 'markdown'-enabled viewer, such as "preview"
within Visual Studio Code.